+++ lib/src/configuration/configuration_ipv4_routing.dart
@@ -1,4 +1,5 @@
 import 'package:cisco_ios_dsl/src/device.dart';
+import 'package:cisco_ios_dsl/src/util/ipv4.dart';
 
 /// The IPv4 SSH configurator for the Configuration scope.
 class ConfigurationIpv4Routing {
@@ -12,6 +13,23 @@ class ConfigurationIpv4Routing {
   void enable() => device
     ..comment('Enable IPv4 routing')
     ..run('ip routing');
+
+  /// Creates a static default route.
+  void createStaticDefaultRoute(IPv4 packetMask, IPv4 destination) => device
+    ..comment('Create a static default IPv4 route')
+    ..comment('  Packet mask IP:      ${packetMask.ip}')
+    ..comment('  Packet mask netmask: ${packetMask.mask}')
+    ..comment('  Destination IP:      ${destination.ip}')
+    ..run('ip route ${packetMask.ip} ${packetMask.mask} ${destination.ip}');
+
+  /// Creates a floating static default route.
+  void createFloatingStaticDefaultRoute(IPv4 packetMask, IPv4 destination, int administrativeDistance) => device
+    ..comment('Create a floating static default IPv4 route')
+    ..comment('  Packet mask IP:          ${packetMask.ip}')
+    ..comment('  Packet mask netmask:     ${packetMask.mask}')
+    ..comment('  Destination IP:          ${destination.ip}')
+    ..comment('  Administrative distance: $administrativeDistance')
+    ..run('ip route ${packetMask.ip} ${packetMask.mask} ${destination.ip} $administrativeDistance');
 
   /// Disables IPv4 routing.
   void disable() => device